refactor(workout-plan): split table renderer into workout-plan-table.js (WP3.4b)
Mechanical move-only extraction of the plan table from the workout-plan.js
monolith into a new workout-plan-table.js module. No behavior, API, DB,
calculation, response, or event-timing change.

Moved: the table renderer (updateWorkoutPlanTable), routine tabs
(updateRoutineTabs / handleRoutineTabClick / initializeRoutineTabs),
muscle-column display transforms (transformMuscleDisplay +
getRelevantIsolatedMuscles / updateMuscleDisplaysInTable / handleViewModeChange),
inline cell edit (makeTableCellEditable / updateExerciseField), per-row cache
and metadata ops (updateRowMetadata / updateCachedExercise — now exported for
the later replacement slice), the swap-button markup, and drag-and-drop
reordering.

To avoid a circular import back into the monolith, the feature callbacks the
renderer and drag behavior depend on — swap, superset toggle, superset action
init, execution-style click, and plan refresh — are injected once at load via
configureWorkoutPlanTable(). The WP3.3 workout-plan-state.js singleton is
imported directly. workout-plan.js re-exports updateWorkoutPlanTable so the
existing E2E harness (mod.updateWorkoutPlanTable) keeps resolving it.

fetchWorkoutPlan() and updateWorkoutPlanUI() stay in workout-plan.js as the
orchestration / page-statistics boundary. WP3.4a (state.js) recorded in
REFACTOR_PLAN.md as delivered by WP3.3.

Gates: Vitest 88 passed; E2E workout-plan / exercise-interactions /
superset-edge-cases / fatigue-context / learned-calibration /
replace-exercise-errors all green (one known learned-calibration timing flake,
green on isolated re-run).
